Add files via upload
Names are now generated from a database, rather than being pulled from a list written into the code.  This utilizes the SQLite module to combine to pull from the lists during the last step when the name is generated.  Occupations are temporarily not available while those databases are being created.