Add files via upload
The GUI is fully operational, although the formatting needs some work.  The DBM has not been integrated into this version yet.  There are a fixes coming, such as alerting the user when a combination returns no results, mainly with alliteration.  The name labels are also not able to be copied, so that will also be fixed.